+++ core/src/main/java/uk/gov/justice/services/core/dispatcher/DispatcherDelegate.java
@@ -7,6 +7,8 @@ import uk.gov.justice.services.core.sender.Sender;
 import uk.gov.justice.services.messaging.Envelope;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
+import java.util.function.Function;
+
 import javax.json.JsonValue;
 
 public class DispatcherDelegate implements Requester, Sender {
@@ -31,20 +33,26 @@ public class DispatcherDelegate implements Requester, Sender {
 
     @Override
     public JsonEnvelope request(final Envelope<?> envelope) {
-        final Envelope<JsonValue> jsonValueEnvelope = envelopePayloadTypeConverter.convert(envelope, JsonValue.class);
-        final JsonEnvelope jsonEnvelope = jsonEnvelopeRepacker.repack(jsonValueEnvelope);
-        final JsonEnvelope response = dispatcher.dispatch(jsonEnvelope);
-
-        requestResponseEnvelopeValidator.validateResponse(response);
-
-        return response;
+        return dispatchAndValidateResponse(envelope);
     }
 
     @Override
     public <T> Envelope<T> request(final Envelope<?> envelope, final Class<T> clazz) {
-        final Envelope<JsonValue> jsonValueEnvelope = envelopePayloadTypeConverter.convert(envelope, JsonValue.class);
-        final JsonEnvelope jsonEnvelope = jsonEnvelopeRepacker.repack(jsonValueEnvelope);
-        final JsonEnvelope response = dispatcher.dispatch(jsonEnvelope);
+        final JsonEnvelope response = dispatchAndValidateResponse(envelope);
+
+        return envelopePayloadTypeConverter.convert(response, clazz);
+    }
+
+    @Override
+    public JsonEnvelope requestAsAdmin(final JsonEnvelope envelope) {
+        final JsonEnvelope response = dispatchAsAdmin().apply(envelope);
+        requestResponseEnvelopeValidator.validateResponse(response);
+        return response;
+    }
+
+    @Override
+    public <T> Envelope<T> requestAsAdmin(final Envelope<?> envelope, final Class<T> clazz) {
+        final JsonEnvelope response = dispatchAsAdmin().compose(convertAndRepackEnvelope()).apply(envelope);
 
         requestResponseEnvelopeValidator.validateResponse(response);
 
@@ -52,29 +60,46 @@ public class DispatcherDelegate implements Requester, Sender {
     }
 
     @Override
-    public JsonEnvelope requestAsAdmin(final JsonEnvelope envelope) {
-        final JsonEnvelope response = dispatchAsAdmin(envelope);
-        requestResponseEnvelopeValidator.validateResponse(response);
-        return response;
-    }
-
-    @Override
     public void send(final Envelope<?> envelope) {
-        final Envelope<JsonValue> jsonValueEnvelope = envelopePayloadTypeConverter.convert(envelope, JsonValue.class);
-        final JsonEnvelope jsonEnvelope = jsonEnvelopeRepacker.repack(jsonValueEnvelope);
+        final JsonEnvelope jsonEnvelope = convertAndRepackEnvelope().apply(envelope);
 
         requestResponseEnvelopeValidator.validateRequest(jsonEnvelope);
 
-        dispatcher.dispatch(jsonEnvelope);
+        dispatch().apply(jsonEnvelope);
     }
 
     @Override
     public void sendAsAdmin(final JsonEnvelope envelope) {
         requestResponseEnvelopeValidator.validateRequest(envelope);
-        dispatchAsAdmin(envelope);
+        dispatchAsAdmin().apply(envelope);
     }
 
-    private JsonEnvelope dispatchAsAdmin(final JsonEnvelope envelope) {
-        return dispatcher.dispatch(systemUserUtil.asEnvelopeWithSystemUserId(envelope));
+    @Override
+    public void sendAsAdmin(final Envelope<?> envelope) {
+        final JsonEnvelope jsonEnvelope = convertAndRepackEnvelope().apply(envelope);
+
+        requestResponseEnvelopeValidator.validateRequest(jsonEnvelope);
+
+        dispatchAsAdmin().apply(jsonEnvelope);
+    }
+
+    private JsonEnvelope dispatchAndValidateResponse(final Envelope<?> envelope) {
+        final JsonEnvelope response = dispatch().compose(convertAndRepackEnvelope()).apply(envelope);
+
+        requestResponseEnvelopeValidator.validateResponse(response);
+        return response;
+    }
+
+    private Function<Envelope<?>, JsonEnvelope> convertAndRepackEnvelope() {
+        return (jsonEnvelope) -> jsonEnvelopeRepacker.repack(
+                envelopePayloadTypeConverter.convert(jsonEnvelope, JsonValue.class));
+    }
+
+    private Function<JsonEnvelope, JsonEnvelope> dispatch() {
+        return (jsonEnvelope) -> dispatcher.dispatch(jsonEnvelope);
+    }
+
+    private Function<JsonEnvelope, JsonEnvelope> dispatchAsAdmin(){
+        return (jsonEnvelope) -> dispatcher.dispatch(systemUserUtil.asEnvelopeWithSystemUserId(jsonEnvelope));
     }
 }